V 0.1.0 Combact armour update. This version wants to rebalance combact armors to give a choice instead of 1 best. There are now 3 classes: tank (netherite), fighter(diamond), leather and studded leather (rogue).
The progression is also altered, the early on armour are now leather and iron, the late game one are diamond, netherite and studded leather. In theory there is no best and so the choice is based on the play stile of each player.

New:
-upgrading tools, used instead of templates to upgrade armours and items
-studded leather armour, an upgraded version of the leather armour with the same stats of iron, is made using leather and netherite
-armour effects, now some armour will give effects. negative effects are given with just one piece while positive ones are given with the full sett
netherite : slowness
leather : speed
studded leather : speed

Changes:
-smithing table is now made with the upgrading template, making it a rarer item
-netherite armour is now made using iron